+++ apps/openmw/mwgui/pickpocketitemmodel.cpp
@@ -7,7 +7,6 @@
 #include "../mwmechanics/creaturestats.hpp"
 #include "../mwmechanics/pickpocket.hpp"
 
-#include "../mwworld/containerstore.hpp"
 #include "../mwworld/class.hpp"
 
 #include "../mwbase/environment.hpp"
@@ -22,7 +21,7 @@ namespace MWGui
     {
         MWWorld::Ptr player = MWMechanics::getPlayer();
         mSourceModel = sourceModel;
-        int chance = player.getClass().getSkill(player, ESM::Skill::Sneak);
+        float chance = player.getClass().getSkill(player, ESM::Skill::Sneak);
 
         mSourceModel->update();
